--- src/Tickets/Seating/app/blockEditor/ticket-is-asc.js
@@ -1,16 +1,0 @@
-import { storeName } from './store';
-import { select } from '@wordpress/data';
-
-/**
- * Filters whether the ticket is ASC.
- *
- * @since TBD
- *
- * @param {boolean} isAsc    Whether ticket is ASC.
- * @param {number}  clientId The ticket ID.
- *
- * @return {boolean} Whether ticket is ASC.
- */
-export const filterTicketIsAsc = (isAsc, clientId) => {
-	return isAsc || !!select(storeName).getTicketSeatType(clientId);
-};
