--- src/Tickets/Seating/app/blockEditor/update-states-seat-type-by-ticket-id.js
@@ -1,6 +1,0 @@
-import { storeName } from './store';
-import { dispatch } from '@wordpress/data';
-
-export const useUpdateStatesSeatTypeByTicketId = (clientId) => {
-	dispatch(storeName).setTicketSeatTypeByPostId(clientId);
-};
